+++ src/main/java/de/fxnm/toolwindow/ToolWindowBase.java
@@ -1,98 +1,75 @@
 package de.fxnm.toolwindow;
 
-import com.intellij.openapi.actionSystem.ActionGroup;
-import com.intellij.openapi.actionSystem.ActionManager;
-import com.intellij.openapi.actionSystem.ActionToolbar;
-import com.intellij.ui.components.JBScrollPane;
-import com.intellij.util.ui.JBUI;
+import com.intellij.openapi.ui.ComboBox;
 
 import java.awt.BorderLayout;
-import java.awt.Component;
+import java.awt.Dimension;
 
-import javax.swing.Box;
-import javax.swing.Icon;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
 
-import de.fxnm.ui.errormessage.ErrorMessagePanel;
+public class ToolWindowBase {
 
-public class ToolWindowBase extends JPanel {
+    private static final ComboBox<String> COMBOBOX = new ComboBox<>(new String[]{"TEST"});
 
-    private final String toolWindowName;
-    private final ErrorMessagePanel errorMessagePanel = new ErrorMessagePanel();
-    private JBScrollPane scrollPane;
-    private JComponent centerJComponent;
+    private final JPanel basePanel;
 
-    public ToolWindowBase(final String toolWindowName) {
-        super(new BorderLayout());
+    private final JComponent northComponent;
+    private final JComponent eastComponent = new JPanel();
+    private final JComponent westComponent;
+    private final JComponent centerComponent;
+    private final JComponent centerTopComponent;
 
-        this.toolWindowName = toolWindowName;
 
-        this.setVisible(true);
+    public ToolWindowBase(final JComponent northComponent,
+                          final JComponent westComponent,
+                          final JComponent centerComponent,
+                          final JComponent centerTopComponent) {
+        this.northComponent = northComponent;
+        this.westComponent = westComponent;
+        this.centerComponent = centerComponent;
+        this.centerTopComponent = centerTopComponent;
+
+        this.basePanel = new JPanel(new BorderLayout());
+
+        this.setSize();
+        this.createPanel();
+
+        this.basePanel.setVisible(true);
     }
 
-    public void createUI(final String sideActionToolbarName,
-                         final JComponent headJComponent,
-                         final JComponent centerJComponent,
-                         final JComponent bottomJComponent) {
+    private void setSize() {
+        final int size = COMBOBOX.getMinimumSize().height;
+        final int height = this.westComponent.getMinimumSize().height + 2;
 
-        this.add(this.createHorizontalToolBar(sideActionToolbarName), BorderLayout.WEST);
-
-        this.add(this.createCenterJComponent(
-                headJComponent,
-                centerJComponent,
-                bottomJComponent
-        ), BorderLayout.CENTER);
-
-        this.setBorder(JBUI.Borders.empty(1));
-        this.setVisible(true);
-
-        this.invalidate();
-        this.repaint();
+        this.setComponentSize(this.northComponent, 0, height);
+        this.setComponentSize(this.eastComponent, size, 0);
+        this.setComponentSize(this.westComponent, size, 0);
     }
 
-    private JComponent createHorizontalToolBar(final String sideActionToolbarName) {
-        final ActionGroup actionGroup = (ActionGroup) ActionManager.getInstance().getAction(sideActionToolbarName);
-        final ActionToolbar actionToolbar = ActionManager.getInstance().createActionToolbar(this.toolWindowName, actionGroup, false);
-        final Box toolBox = Box.createHorizontalBox();
-        toolBox.add(actionToolbar.getComponent());
-        return toolBox;
+    private void setComponentSize(final JComponent component, final int width, final int height) {
+        final Dimension dimension = new Dimension(width, height);
+        component.setMinimumSize(dimension);
+        component.setPreferredSize(dimension);
     }
 
-    private JComponent createCenterJComponent(final JComponent headJComponent,
-                                              final JComponent centerJComponent,
-                                              final JComponent bottomJComponent) {
-        final JPanel toolPanel = new JPanel(new BorderLayout());
-        final JPanel topToolPanel = new JPanel(new BorderLayout());
 
-        topToolPanel.add(headJComponent, BorderLayout.NORTH);
-        topToolPanel.add(this.errorMessagePanel, BorderLayout.SOUTH);
+    private void createPanel() {
+        final JPanel innerCenterPanel = new JPanel(new BorderLayout());
+        innerCenterPanel.add(this.centerTopComponent, BorderLayout.NORTH);
+        innerCenterPanel.add(this.centerComponent, BorderLayout.CENTER);
 
-        toolPanel.add(topToolPanel, BorderLayout.NORTH);
-        this.scrollPane = new JBScrollPane();
-        this.centerJComponent = centerJComponent;
-        this.scrollPane.setViewportView(centerJComponent);
 
-        toolPanel.add(this.scrollPane, BorderLayout.CENTER);
-        toolPanel.add(bottomJComponent, BorderLayout.SOUTH);
+        final JPanel innerPanel = new JPanel(new BorderLayout());
+        innerPanel.add(this.northComponent, BorderLayout.NORTH);
+        innerPanel.add(innerCenterPanel, BorderLayout.CENTER);
 
-        return toolPanel;
+        this.basePanel.add(this.eastComponent, BorderLayout.EAST);
+        this.basePanel.add(innerPanel, BorderLayout.CENTER);
+        this.basePanel.add(this.westComponent, BorderLayout.WEST);
     }
 
-    public void setMessage(final Icon icon, final Boolean autoRemove, final String message) {
-        this.errorMessagePanel.setErrorMessages(icon, autoRemove, message);
-        this.invalidate();
-        this.repaint();
-        this.setVisible(true);
-    }
-
-    public void changeViewport(final Component component) {
-        this.scrollPane.setViewportView(component);
-        this.repaint();
-    }
-
-    public void restoreViewport() {
-        this.scrollPane.setViewportView(this.centerJComponent);
-        this.repaint();
+    public JPanel getBasePanel() {
+        return this.basePanel;
     }
 }