+++ src/main/java/icons/PluginIcons.java
@@ -9,7 +9,7 @@ public interface PluginIcons {
     Icon PLUGIN_ICON = AllIcons.Modules.TestResourcesRoot;
 
     Icon LOGIN_GREEN = IconLoader.getIcon("/de/fxnm/icons/login/status/LoginGreen.svg", PluginIcons.class);
-    Icon LOGIN_RED = IconLoader.getIcon("/de/fxnm/icons/login/status/LoginRed.svg",PluginIcons.class);
+    Icon LOGIN_RED = IconLoader.getIcon("/de/fxnm/icons/login/status/LoginRed.svg", PluginIcons.class);
 
     Icon STATUS_SUCCESS = AllIcons.Actions.Commit;
     Icon STATUS_ERROR = AllIcons.Vcs.Remove;